% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_perform_wilcox_rank_sum_test_evaluations}
\alias{analyse_perform_wilcox_rank_sum_test_evaluations}
\title{FUNCTION: analyse_perform_wilcox_rank_sum_test_evaluations}
\usage{
analyse_perform_wilcox_rank_sum_test_evaluations(s, v)
}
\description{
Perform a single Wilcoxon rank-sum test for two input data sets for evaluations
}
