% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\alias{analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\title{FUNCTION: analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\usage{
analyse_vargha_delaney_effect_size_except_testgnerationtiming(d, e)
}
\description{
Perform an effect size analysis of the coverage, mutation score, and evaluations for the AVM and Directed Random techniques.
}
