% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_vargha_delaney_effect_size_thresholding}
\alias{analyse_vargha_delaney_effect_size_thresholding}
\title{FUNCTION: analyse_vargha_delaney_effect_size_thresholding}
\usage{
analyse_vargha_delaney_effect_size_thresholding(d)
}
\description{
Perform an effect size analysis of the timings for the AVM and DR mutation analysis techniques.
By default, specify that the calculation should use the effectsize_accurate function for calculating the statistic.
But, investigate the effect size with different levels of thresholding, to establish sensitivity.
}
