% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutanttiming}
\alias{collect_mutanttiming}
\title{FUNCTION: collect_mutanttiming}
\usage{
collect_mutanttiming()
}
\value{
A Data Frame of mutants
}
\description{
Preform binding for all split files for mutatnttimng.
}
