% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_mutationanalysistime}
\alias{collect_mutationanalysistime}
\title{FUNCTION: collect_mutationanalysistime}
\usage{
collect_mutationanalysistime()
}
\description{
Preform binding for all split files for mutationanalysistime.
}
