% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{generate_15_schmea_mutants}
\alias{generate_15_schmea_mutants}
\title{FUNCTION: generate_15_schmea_mutanttiming}
\usage{
generate_15_schmea_mutants(d)
}
\arguments{
\item{d}{A data frame of mutants}
}
\value{
A data frame with 15 schema randomly picked
}
\description{
Gets a 15 schemas from the dataset of mutants file, or mutanttiming.dat.
}
