% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{generate_15_schmea_mutanttiming}
\alias{generate_15_schmea_mutanttiming}
\title{FUNCTION: generate_15_schmea_mutanttiming}
\usage{
generate_15_schmea_mutanttiming(d)
}
\description{
gets a 15 schemas from the dataset of mutanttiming file.
}
