% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{generate_15_schmea_mutation_analysis}
\alias{generate_15_schmea_mutation_analysis}
\title{FUNCTION: generate_15_schmea_mutation_analysis}
\usage{
generate_15_schmea_mutation_analysis(d)
}
\description{
gets a 15 schemas from the dataset of mutantionanalysistime file.
}
