% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{ordering_mutants_per_schema}
\alias{ordering_mutants_per_schema}
\title{FUNCTION: ordering_mutants_per_schema}
\usage{
ordering_mutants_per_schema(d)
}
\description{
generates a a data frame ready with runs for mutant timing file per schema.
}
