% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_heatmap_mutanttiming}
\alias{plot_heatmap_mutanttiming}
\title{FUNCTION: plot_heatmap_mutanttiming}
\usage{
plot_heatmap_mutanttiming(d)
}
\arguments{
\item{d}{Data frame of analysis}
}
\description{
Perform a heating plot for each case
File saved in plots/
}
