% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_heatmap_mutanttiming_allinone}
\alias{plot_heatmap_mutanttiming_allinone}
\title{FUNCTION: plot_heatmap_mutanttiming_allinone}
\usage{
plot_heatmap_mutanttiming_allinone(d)
}
\arguments{
\item{d}{Data frame of analysis}
}
\value{
A Plot
}
\description{
Perform a heating plot for all cases
File saved in plots/
}
