% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_heatmap_mutanttiming_allinone_dbms}
\alias{plot_heatmap_mutanttiming_allinone_dbms}
\title{FUNCTION: plot_heatmap_mutanttiming_allinone_dbms}
\usage{
plot_heatmap_mutanttiming_allinone_dbms(d)
}
\arguments{
\item{d}{Data frame of analysis}
}
\value{
A Plot
}
\description{
Perform a heating plot for all cases with dbms
File saved in plots/
}
