% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_overall}
\alias{plot_overall}
\title{FUNCTION: plot_overall}
\usage{
plot_overall(d)
}
\arguments{
\item{d}{Data frame of analysis}
}
\description{
Perform a plotting for coverage, mutation score, evaluations and testgenerationtime
File saved in plots/
}
