% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{preform_averaging_of_samples}
\alias{preform_averaging_of_samples}
\title{FUNCTION: preform_averaging_of_samples}
\usage{
preform_averaging_of_samples(d)
}
\arguments{
\item{d}{A data frame of analysis}
}
\value{
Averaged samples for all runs
}
\description{
Average out all schemas for each data generators and DBMS for each run.
}
