% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{preform_sum_of_samples}
\alias{preform_sum_of_samples}
\title{FUNCTION: preform_sum_of_samples}
\usage{
preform_sum_of_samples(d)
}
\arguments{
\item{d}{A data frame of analysis}
}
\value{
Total results (summtion)
}
\description{
Total of all results per schema for each data generators and DBMS for each run.
}
