% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_coverage}
\alias{siginificant_coverage}
\title{FUNCTION: siginificant_coverage}
\usage{
siginificant_coverage(d, rtrn = "tex", m = "median")
}
\arguments{
\item{d}{Data frame of analysis}

\item{rtrn}{Latex (tex) or a data frame (data)}

\item{m}{Results shown as median or mean}
}
\value{
A A12 effect size and U-test of coverages compared pair wise
}
\description{
Generates a latex table or a data frame for coverage table with effect size and U test.
}
