% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_mutant_operators}
\alias{siginificant_mutant_operators}
\title{FUNCTION: siginificant_mutant_operators}
\usage{
siginificant_mutant_operators(d)
}
\description{
generates a latex table for mutation operators table with effect size and U test.
}
