% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_mutant_operators_fixed}
\alias{siginificant_mutant_operators_fixed}
\title{FUNCTION: siginificant_mutant_operators_fixed}
\usage{
siginificant_mutant_operators_fixed(d, rtrn = "tex")
}
\description{
generates a latex table for mutation operators table with effect size and U test.
}
