% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_mutation_score}
\alias{siginificant_mutation_score}
\title{FUNCTION: siginificant_mutation_score}
\usage{
siginificant_mutation_score(d, rtrn = "tex", m = "median")
}
\arguments{
\item{d}{Data frame of mutants}

\item{rtrn}{Latex (tex) or a data frame (data)}

\item{m}{Results shown as median or mean}
}
\value{
A A12 effect size and U-test of mutation score per schema compared pair wise
}
\description{
Generates a latex table or data frame for mutation score per schema table with effect size and U test.
}
