% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{testtimegeneration_table_bolding}
\alias{testtimegeneration_table_bolding}
\title{FUNCTION: testtimegeneration_table_bolding}
\usage{
testtimegeneration_table_bolding(d)
}
\description{
Perform a latex bolding and returning a dataframe
}
