% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_execution_times_for_threshold}
\alias{transform_execution_times_for_threshold}
\title{FUNCTION: transform_execution_times_for_threshold}
\usage{
transform_execution_times_for_threshold(d, t)
}
\arguments{
\item{d}{Data frame of mutants}

\item{t}{A threshold time (ms)}
}
\value{
A data frame of a transforemd test generation timing
}
\description{
Transform the timing values so that they do not include any below a specified number of ms.
Include rounding down
}
