% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visual_grid_ten_schemas}
\alias{visual_grid_ten_schemas}
\title{FUNCTION: visual_grid_ten_schemas}
\usage{
visual_grid_ten_schemas(mutationanalysistiming)
}
\description{
generates a plot grid of top 10 schemas
}
