% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_perform_wilcox_rank_sum_test_evaluations}
\alias{analyse_perform_wilcox_rank_sum_test_evaluations}
\title{FUNCTION: analyse_perform_wilcox_rank_sum_test_evaluations}
\usage{
analyse_perform_wilcox_rank_sum_test_evaluations(s, v)
}
\arguments{
\item{s}{Sample 1}

\item{v}{Sample 2}
}
\value{
A textual results
}
\description{
Perform a single Wilcoxon rank-sum test for two evaluation results samples
}
