% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\alias{analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\title{FUNCTION: analyse_vargha_delaney_effect_size_except_testgnerationtiming}
\usage{
analyse_vargha_delaney_effect_size_except_testgnerationtiming(d, e)
}
\arguments{
\item{d}{A data frame of analysis}

\item{e}{The effect size method}
}
\value{
Data frame of coverage, mutation score, and evaluations A12 effect size results (Only AVM-D and DR)
}
\description{
Perform an effect size analysis of the coverage, mutation score, and evaluations for the AVM-D and Directed Random techniques.
}
