% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_vargha_delaney_effect_size_testgenerationtiming}
\alias{analyse_vargha_delaney_effect_size_testgenerationtiming}
\title{FUNCTION: analyse_vargha_delaney_effect_size_testgenerationtiming}
\usage{
analyse_vargha_delaney_effect_size_testgenerationtiming(d, e)
}
\description{
Perform an effect size analysis of the test generation timings for the AVM and Directed Random mutation analysis techniques.
}
