% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_analyse.R
\name{analyse_wilcox_rank_sum_test}
\alias{analyse_wilcox_rank_sum_test}
\title{FUNCTION: analyse_wilcox_rank_sum_test}
\usage{
analyse_wilcox_rank_sum_test(d)
}
\description{
Perform a statistical analysis of the timings for the AVM and directed random mutation analysis techniques. Right now, this method
simply returns textual output; it would be much better if it actually returned a single data frame.
}
