% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{generate_15_schmea_analysis}
\alias{generate_15_schmea_analysis}
\title{FUNCTION: generate_15_schmea_mutation_analysis}
\usage{
generate_15_schmea_analysis(d)
}
\arguments{
\item{d}{A data frame of analysis}
}
\value{
A data frame with 15 schema randomly picked
}
\description{
Gets a 15 schemas from the dataset of analysis file, or mutantionanalysistime.dat.
}
