% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_coverage}
\alias{siginificant_coverage}
\title{FUNCTION: siginificant_coverage}
\usage{
siginificant_coverage(d, rtrn = "tex")
}
\description{
generates a latex table for coverage table with effect size and U test.
}
