% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_mutant_operators}
\alias{siginificant_mutant_operators}
\title{FUNCTION: siginificant_mutant_operators}
\usage{
siginificant_mutant_operators(d, rtrn = "tex", m = "median")
}
\arguments{
\item{d}{Data frame of mutants}

\item{rtrn}{Latex (tex) or a data frame (data)}

\item{m}{Results shown as median or mean}
}
\value{
A A12 effect size and U-test of mutation score per operator compared pair wise
}
\description{
Generates a latex table or data frame for mutation operators table with effect size and U test.
}
