% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_mutation_score}
\alias{siginificant_mutation_score}
\title{FUNCTION: siginificant_mutation_score}
\usage{
siginificant_mutation_score(d)
}
\description{
generates a latex table for mutation score per schema table with effect size and U test.
}
