% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_generator.R
\name{siginificant_timing}
\alias{siginificant_timing}
\title{FUNCTION: siginificant_timing}
\usage{
siginificant_timing(d)
}
\description{
generates a latex table for test generation timing table with effect size and U test.
}
